CCT 1.0.11 Release Notes Features and changes: Collapsed the envelope and common modules down into calibration and moved the envelope tool from a standalone app to a sub-application in the tool menu. Updated the site and measurement Mw fitting algorithm to re-weight the fits by the standard error of the frequency band population in an effort to improve the fitting robustness and get better uncertainty estimates
Autopicking updates:
Now a two-step iterative process where we first take the SNR end pick, compute shapes, measure the deviation of each fit against the frequency model, pick the divergence point, and then re-measure with the new endpoint as a variable.
See the new Parameters->Shape Constraints->Fit Length parameter to adjust the tradeoff of envelope length vs. misfit. The range is [0,1] and is the ratio of weight given to the length where 0 dis-regards length and 1 disregards misfit.

Constraint changes:
Maximum stress changed from 10 to 100 MPa
Widened the path constraints in Joint1DPathCorrection to admit much higher Q values and allow longer crossover distances to support additional regions and higher frequency bands.

Adjusted the spectra truncation feature that is enabled when you have both a GT source spectra and Mw only reference events to use the MDAC sigma/psi-based falloff to compute the truncation corner frequency rather than an arbitrarily low value. This allows you to control how many bands on average are used for the Mw only events as site term constraints. This change was introduced for situations where the GT spectra events might not contain sufficient measurements below the corner frequency.

Bugfixes:
Addressed three potential NPEs and nine potential resources handle leaks
Updated all the entry points to set the default locale to English to prevent numbers being converted from using periods in countries that use comma separators. This is a temporary workaround to prevent functional errors related to envelope creation and reading.
Updates for the latest release versions of all dependencies.

+++ externals/src/main/java/llnl/gnem/core/util/randomNumbers/BaseRandomAlgorithm.java
@@ -40,7 +40,7 @@ public abstract class BaseRandomAlgorithm implements RandomAlgorithm {
         long max = Integer.MAX_VALUE;
         long inRange = max - min;
         long value = this.nextInt();
-        double numerator = (value - min) * outRange;
+        double numerator = (value - min) * (double) outRange;
         long result = lower + Math.round(numerator / inRange);
         return (int) result;
     }
